# resizable_memory: use mprotect instead of mmap to resize on linux