Uses `mprotect()` instead of `mmap()` to add the additional pages. I'm amazed how much faster this is. Unmapping is supported too now and should allow the OS to reclaim unused pages.

### Before

Mapping 1048576 bytes in increments of 1009 bytes

| relative |               ns/op |                op/s |    err% |          ins/op |         bra/op |   miss% |     total | benchmark
|---------:|--------------------:|--------------------:|--------:|----------------:|---------------:|--------:|----------:|:----------
|   100.0% |          190,850.05 |            5,239.72 |    0.3% |       17,428.97 |       4,422.01 |   12.0% |     12.12 | `mmap+mprotect`
|    17.3% |        1,106,018.39 |              904.14 |    0.1% |       34,734.03 |       9,592.02 |    5.6% |     12.07 | `resizable_memory`

Mapping 1048576 bytes in increments of 4096 bytes

| relative |               ns/op |                op/s |    err% |          ins/op |         bra/op |   miss% |     total | benchmark
|---------:|--------------------:|--------------------:|--------:|----------------:|---------------:|--------:|----------:|:----------
|   100.0% |          190,977.82 |            5,236.21 |    0.4% |        5,521.96 |       1,838.01 |   28.2% |     12.06 | `mmap+mprotect`
|    17.3% |        1,104,988.38 |              904.99 |    0.2% |       16,191.04 |       4,113.02 |   12.6% |     12.09 | `resizable_memory`

### After

Mapping 1048576 bytes in increments of 1009 bytes

| relative |               ns/op |                op/s |    err% |          ins/op |         bra/op |   miss% |     total | benchmark
|---------:|--------------------:|--------------------:|--------:|----------------:|---------------:|--------:|----------:|:----------
|   100.0% |          192,256.29 |            5,201.39 |    0.1% |       17,428.97 |       4,422.01 |   12.0% |     12.09 | `mmap+mprotect`
|   106.5% |          180,528.21 |            5,539.30 |    0.1% |       28,085.02 |       7,534.02 |    7.1% |     12.11 | `resizable_memory`

Mapping 1048576 bytes in increments of 4096 bytes

| relative |               ns/op |                op/s |    err% |          ins/op |         bra/op |   miss% |     total | benchmark
|---------:|--------------------:|--------------------:|--------:|----------------:|---------------:|--------:|----------:|:----------
|   100.0% |          191,424.31 |            5,224.00 |    0.2% |        5,521.98 |       1,838.01 |   28.2% |     12.09 | `mmap+mprotect`
|   106.4% |          179,915.29 |            5,558.17 |    0.4% |        9,806.02 |       2,837.01 |   18.2% |     12.13 | `resizable_memory`

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Bug Fixes**
  - Improved memory resizing reliability for mapped files on Linux, with better handling of size limits and memory protection.
- **Performance**
  - Enhanced efficiency when resizing mapped memory regions by incrementally adjusting protections instead of remapping the entire range.
- **Tests**
  - Added new Linux-only tests to verify page residency behavior after decommitting memory.
  - Refactored existing memory resize tests for better isolation and clarity.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->