@pknowles 2025-08-04
*for reference, `MADV_DONTNEED` does slow things down a little. Not nearly as much as a full remap with `mmap()`. For the moment, keeping the size as the actual mapped range unlike e.g. std::vector with a separate capacity().